Extract smaller theme components for typography
