+++ src/theme/typography/brandFont.js
@@ -1,0 +1,1 @@
+export default 'Motiva Sans';
