+++ src/theme/typography/fontWeights.js
@@ -1,0 +1,5 @@
+export default {
+  regular: 400,
+  medium: 500,
+  bold: 700
+};
