+++ src/theme/typography/fonts.js
@@ -1,0 +1,6 @@
+import brandFont from './brandFont';
+
+export default {
+  default: brandFont,
+  icon: 'Arbor'
+};
